timife/feat/db relationship

+++ categories/src/main/java/com/timife/repositories/SectionRepository.java
@@ -12,5 +12,5 @@ public interface SectionRepository extends JpaRepository<Section, Long> {
     @Query(
             value = "SELECT * FROM section u WHERE u.name = ?1 AND u.category_id = ?2",
             nativeQuery = true)
-    Section findSectionByNameAndCategory(String name, Long categoryId);
+    Section findSectionByNameAndCategory(String name, Integer categoryId);
 }
